// InterLex Editor Tool/Interlex.Angular.App/src/app/models/case-editor.model.ts
export interface ICaseMetaContentResponse {
  content: string;
  editable: boolean;
}

export interface IExpertContentResponse extends ICaseMetaContentResponse {
  files: IFile[];
}
// ...
export interface IExpertMaterial {
  title?: string;
  category?: INomenclature;
  journal?: string;
  issue?: string;
  pages?: string;
  author?: string[];
  corporateAuthor?: string[];
  publisher?: string[];
  publicationDate?: Date;
  publicationYear?: string;
  sources?: ISource[];
  abstract?: string;
  keywords?: string[];
  issn?: string[];
  isbn?: string[];
  doi?: string[];
  languages?: string[];
  formats?: string[];
  license?: ILicense;
  text?: string;
  // add others later

}

export interface ISource {
  name?: string;
  url?: string;
}

export interface ILicense {
  label?: string;
  value?: string;
  title?: string;
}

export interface IFile {
  id?: string;
  base64Content?: string;
  filename?: string;
  mimeType?: string;
  language?: LanguageModel;
}
// ...
export interface INomenclature {
  label: string;
  code: string;
}